run game logic in a separate thread from rendering
Very messy implementation, and I think it's worth exploring an
approach without any shared state and locking with RwLock. Instead,
try using a queue that we share between threads, and fill up on state
update. (That queue might have length 0 though, as we always want to
draw the latest state.)

closes #11