clamp intended head/arms angle automatically
Example: I say "turn the head 360 to the left", and my head is
currently pointing forward, that is, angled at 0 relative to my body.
Now the intent will not be -2pi, but rather the maximum achievable
relative position, which is -pi/2.0 in this case.

This simplifies some calculations with the "remaining turn angle" for
both arms and head.