uniquify players
Since it should be possible to add the "same" player (implementation)
multiple times, for instance to battle against itself, the things we use
as keys (meta data for now) need to be unique.
A little hacky, but a simple counter works for now (but should be
improved later, most likely).